feat(ui): depth-first walker for nested archive extraction (Phase 53 task 2)
walk_archive descends into archives nested inside archives, each becoming a
sub-gallery under the gallery holding it, bounded by RecursionBudget.

A recursive plan cannot be a flat placement list: each placement belongs to
a DIFFERENT archive buffer, so an entry index back into "the" entry list is
meaningless once nesting starts. The walker therefore emits as it goes,
through hooks, the way Phase 50's sinks already work — and the archive
backends are injected rather than linked, so the recursion is tested with
fakes and this module never pulls in miniz or libarchive.

Depth-first is load-bearing, not stylistic: only the archives along the
current root->leaf path are alive at once, which is exactly what the
live-bytes limit polices. breadth-first would hold every archive at a level.

Behaviour pinned by tests: several levels of nesting; the depth cap failing
only its branch while sibling media still imports; CBZ never descended
(a comic archive is a flat run of pages); an unreadable nested archive
tallied without disturbing siblings; cancellation honoured between entries;
and two sibling folders both holding "bonus.zip" not collapsing onto one
sub-gallery.

The magic-byte re-check earns its place here: nested candidates are chosen
by extension at plan time, and detect_archive_kind gets the final say once
the bytes exist. My first fake omitted the magic and the walker correctly
refused to descend — the fixture was wrong, not the code.

Also fixes -Wconversion errors in the harvested test_spanned_zip.cpp that
only the --asan configuration's stricter warnings caught; the plain Debug
build compiled them fine.

17 tests; 1392 -> 1409, 0 failed under both scripts/test.sh and --asan.

+++ tests/ui/test_spanned_zip.cpp
@@ -233,13 +233,13 @@ TEST(eocd_disk_fields_rewritten_to_zero) {
     REQUIRE(eocd_pos + 22 <= result.merged.size());
 
     // Check disk_num field (offset +4 in EOCD)
-    uint16_t disk_num = result.merged[eocd_pos + 4] |
-                        (result.merged[eocd_pos + 5] << 8);
+    uint16_t disk_num = static_cast<uint16_t>(
+        result.merged[eocd_pos + 4] | (result.merged[eocd_pos + 5] << 8));
     CHECK_EQ(disk_num, 0);
 
     // Check disk_with_cd field (offset +6)
-    uint16_t disk_with_cd = result.merged[eocd_pos + 6] |
-                            (result.merged[eocd_pos + 7] << 8);
+    uint16_t disk_with_cd = static_cast<uint16_t>(
+        result.merged[eocd_pos + 6] | (result.merged[eocd_pos + 7] << 8));
     CHECK_EQ(disk_with_cd, 0);
 }
 
@@ -279,8 +279,8 @@ TEST(cd_entries_disk_start_set_to_zero) {
     REQUIRE(cd_pos + 46 <= result.merged.size());
 
     // Check disk_start field (offset +34 in CD entry)
-    uint16_t disk_start = result.merged[cd_pos + 34] |
-                          (result.merged[cd_pos + 35] << 8);
+    uint16_t disk_start = static_cast<uint16_t>(
+        result.merged[cd_pos + 34] | (result.merged[cd_pos + 35] << 8));
     CHECK_EQ(disk_start, 0);
 }
